hotfix(compile): split ssf_module with mmengine and torch scatter. * so that users can run other models even without engine or scatter. * move two packages into README for the extra package at least for now.... in case people struggling with setup env again.
docs update ssf model link.
