+++ assets/cuda/mmcv/__init__.py
@@ -1,9 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .voxelize import Voxelization, voxelization
 from .scatter_points import DynamicScatter, dynamic_scatter
-from .conv import *
-from .norm import *
-from .plugin import *
 
 __all__ = [
     'Voxelization', 'voxelization',
