+++ assets/cuda/mmdet/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .conv import *
+from .norm import *
+from .plugin import *
+from .resnet import BasicBlock, Bottleneck
