--- assets/mmdet/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-from .res_layer import *
-from .resnet import *
-from .typing_utils import *
-
-__all__ = []
