+++ src/models/__init__.py
@@ -19,7 +19,7 @@ try:
     from .flow4d import Flow4D
 except ImportError as e:
     print("\033[93m--- WARNING [model]: Model with SparseConv is not imported, as it requires spconv lib which is not installed.")
-    print(f"\033[91m--- Detail error message\033[0m: {e}")
+    print(f"Detail error message\033[0m: {e}. Just ignore this error if code runs without these models.")
 
 # following need install extra package:
 # * pip install torch_scatter mmengine-lite
@@ -27,4 +27,4 @@ try:
     from .ssf import SSF
 except ImportError as e:
     print("\033[93m--- WARNING [model]: Model with torch scatter is not imported, as it requires some lib which is not installed.")
-    print(f"\033[91m--- Detail error message\033[0m: {e}")
+    print(f"Detail error message\033[0m: {e}. Just ignore this warning if code runs without these models.")
